pyproject.toml
[project]
name = "translator-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chainlit>=2.5.5",
    "google-generativeai>=0.8.5",
    "python-dotenv>=1.1.0",
]
